Renames files to be plural/non-plural to match our own convention.
