+++ app.js
@@ -19,7 +19,7 @@ app.use(bodyParser.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
-var rentals = require("./routes/rentals_routes");
+var rentals = require("./routes/rental_routes");
 app.use("/rentals", rentals);
 
 var zomgRoutes = require('./routes/zomgRoutes');
