+++ routes/rental_routes.js
